+++ hcore/hstring.cxx
@@ -63,7 +63,7 @@ int long kmpsearch( char const* const, int long, char const* const, int long );
 
 #undef D_WHITE_SPACE
 #define D_WHITE_SPACE "\a\b \t\v\f\r\n"
-HCharacterClass const _whiteSpace_ = HCharacterClass( D_WHITE_SPACE, sizeof ( D_WHITE_SPACE ) - 1 );
+HCharacterClass const _whiteSpace_ = HCharacterClass( D_WHITE_SPACE, static_cast<int>( sizeof ( D_WHITE_SPACE ) - 1 ) );
 #undef D_WHITE_SPACE
 #undef D_LOWER_CASE_LETTER
 #undef D_UPPER_CASE_LETTER
@@ -73,11 +73,11 @@ HCharacterClass const _whiteSpace_ = HCharacterClass( D_WHITE_SPACE, sizeof ( D_
 #define D_LETTER D_LOWER_CASE_LETTER D_UPPER_CASE_LETTER
 #undef D_DIGIT
 #define D_DIGIT "0123456789"
-HCharacterClass const _digit_ = HCharacterClass( D_DIGIT, sizeof ( D_DIGIT ) - 1 );
-HCharacterClass const _letter_ = HCharacterClass( D_LETTER, sizeof ( D_LETTER ) - 1 );
-HCharacterClass const _lowerCaseLetter_ = HCharacterClass( D_LOWER_CASE_LETTER, sizeof ( D_LOWER_CASE_LETTER ) - 1 );
-HCharacterClass const _upperCaseLetter_ = HCharacterClass( D_UPPER_CASE_LETTER, sizeof ( D_UPPER_CASE_LETTER ) - 1 );
-HCharacterClass const _word_ = HCharacterClass( D_LETTER D_DIGIT "_", sizeof ( D_LETTER D_DIGIT "_" ) - 1 );
+HCharacterClass const _digit_ = HCharacterClass( D_DIGIT, static_cast<int>( sizeof ( D_DIGIT ) - 1 ) );
+HCharacterClass const _letter_ = HCharacterClass( D_LETTER, static_cast<int>( sizeof ( D_LETTER ) - 1 ) );
+HCharacterClass const _lowerCaseLetter_ = HCharacterClass( D_LOWER_CASE_LETTER, static_cast<int>( sizeof ( D_LOWER_CASE_LETTER ) - 1 ) );
+HCharacterClass const _upperCaseLetter_ = HCharacterClass( D_UPPER_CASE_LETTER, static_cast<int>( sizeof ( D_UPPER_CASE_LETTER ) - 1 ) );
+HCharacterClass const _word_ = HCharacterClass( D_LETTER D_DIGIT "_", static_cast<int>( sizeof ( D_LETTER D_DIGIT "_" ) - 1 ) );
 #undef D_DIGIT
 #undef D_LETTER
 #undef D_LOWER_CASE_LETTER
