+++ hcore/hudpsocket.cxx
@@ -108,7 +108,7 @@ void HUDPSocket::bind( int port_, ip_t ip_ ) {
 	address.sin_family = AF_INET;
 	address.sin_port = fwd_htons( static_cast<int short unsigned>( port_ ) );
 	address.sin_addr.s_addr = ip_.raw();
-	M_ENSURE_EX( ( ::bind( _fileDescriptor, reinterpret_cast<sockaddr*>( &address ), sizeof ( address ) ) == 0 ), resolver::ip_to_string( ip_ ) + ":" + port_ );
+	M_ENSURE_EX( ( ::bind( _fileDescriptor, reinterpret_cast<sockaddr*>( &address ), static_cast<int>( sizeof ( address ) ) ) == 0 ), resolver::ip_to_string( ip_ ) + ":" + port_ );
 	return;
 	M_EPILOG
 }
@@ -119,7 +119,7 @@ void HUDPSocket::send_to( ip_t dest_, int port_, void const* data_, int long siz
 	address.sin_family = AF_INET;
 	address.sin_port = fwd_htons( static_cast<int short unsigned>( port_ ) );
 	address.sin_addr.s_addr = dest_.raw();
-	M_ENSURE( ::sendto( _fileDescriptor, data_, size_, 0, reinterpret_cast<sockaddr*>( &address ), sizeof ( address ) ) == size_ );
+	M_ENSURE( ::sendto( _fileDescriptor, data_, size_, 0, reinterpret_cast<sockaddr*>( &address ), static_cast<int>( sizeof ( address ) ) ) == size_ );
 	return;
 	M_EPILOG
 }
@@ -127,7 +127,7 @@ void HUDPSocket::send_to( ip_t dest_, int port_, void const* data_, int long siz
 int long HUDPSocket::receive( ODatagram& datagram_ ) {
 	M_PROLOG
 	sockaddr_in address;
-	socklen_t len( sizeof ( address ) );
+	socklen_t len( static_cast<int>( sizeof ( address ) ) );
 	int long nRead( ::recvfrom( _fileDescriptor, datagram_._data.raw(), datagram_._size, 0, reinterpret_cast<sockaddr*>( &address ), &len ) );
 	datagram_._ip = ip_t( address.sin_addr.s_addr );
 	datagram_._port = fwd_ntohs( address.sin_port );
